requirements.txt
python-dotenv

Django==4.1
django-cors-headers
django-ckeditor==6.4.2
django-import-export==2.8.0
djangorestframework==3.14.0
DateTime==4.5

social-auth-app-django==5.4.1
whitenoise==6.4.0
requests==2.28.1

six
psycopg2-binary